+++ integration/app/examples/count/count.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { NgxsModule } from '@ngxs/store';
+import { NgxsDataUtilsModule } from '@ngxs-labs/data/utils';
 
 import { CountComponent } from './count.component';
 import { CountState } from './count.state';
 import { FormsModule } from '@angular/forms';
 import { CountSubState } from './count-sub.state';
-import { NgxsDataUtilsModule } from '@ngxs-labs/data';
 
 @NgModule({
     declarations: [CountComponent],