+++ src/lib/utils/public-api.ts
@@ -1,0 +1,2 @@
+export { NgxsDataUtilsModule } from './external/modules/ngxs-data-utils.module';
+export { NgxsDataMutablePipe } from './external/modules/pipes/ngxs-data-mutable.pipe';